refactor: 컨설팅 게시글 임시저장 코드 정리
